added other `conversation` APIs
